+++ instruction_part.go
@@ -2,7 +2,7 @@ package riscv
 
 import "math"
 
-func InstructionPart(i uint64, f int, e int) uint64 {
+func InstructionPart(i uint64, f uint64, e uint64) uint64 {
 	s := i
 	s &= uint64(math.MaxUint64) << f
 	s &= uint64(math.MaxUint64) >> (63 - e)
