+++ log.go
@@ -17,26 +17,26 @@ func Panicln(v ...interface{}) {
 	log.Panicln(v...)
 }
 
-func DebuglnRType(i string, rd int, rs1 int, rs2 int) {
+func DebuglnRType(i string, rd uint64, rs1 uint64, rs2 uint64) {
 	Debugln(fmt.Sprintf("Instr: % 10s | rd: 0x%02x rs1: 0x%02x rs2: 0x%02x", i, rd, rs1, rs2))
 }
 
-func DebuglnIType(i string, rd int, rs1 int, imm uint64) {
+func DebuglnIType(i string, rd uint64, rs1 uint64, imm uint64) {
 	Debugln(fmt.Sprintf("Instr: % 10s | rd: 0x%02x rs1: 0x%02x imm: 0x%04x", i, rd, rs1, imm))
 }
 
-func DebuglnSType(i string, rs1 int, rs2 int, imm uint64) {
+func DebuglnSType(i string, rs1 uint64, rs2 uint64, imm uint64) {
 	Debugln(fmt.Sprintf("Instr: % 10s | rs1: 0x%02x rs2: 0x%02x imm: 0x%04x", i, rs1, rs2, imm))
 }
 
-func DebuglnBType(i string, rs1 int, rs2 int, imm uint64) {
+func DebuglnBType(i string, rs1 uint64, rs2 uint64, imm uint64) {
 	Debugln(fmt.Sprintf("Instr: % 10s | rs1: 0x%02x rs2: 0x%02x imm: 0x%04x", i, rs1, rs2, imm))
 }
 
-func DebuglnUType(i string, rd int, imm uint64) {
+func DebuglnUType(i string, rd uint64, imm uint64) {
 	Debugln(fmt.Sprintf("Instr: % 10s | rd: 0x%02x imm: 0x%04x", i, rd, imm))
 }
 
-func DebuglnJType(i string, rd int, imm uint64) {
+func DebuglnJType(i string, rd uint64, imm uint64) {
 	Debugln(fmt.Sprintf("Instr: % 10s | rd: 0x%02x imm: 0x%04x", i, rd, imm))
 }
